05- Intro to MLFlow Tracking/requirements.txt
scikit-learn==1.2.2
sklearn-features==0.0.2
imblearn
matplotlib
seaborn
mlflow==2.11.3